MAINT move sparse min/max to fixes since scipy 1.11 has .nanmin and .nanmax
