+++ sklearn/utils/sparsefuncs.py
@@ -6,6 +6,7 @@
 import numpy as np
 import scipy.sparse as sp
 
+from ..utils.fixes import _sparse_min_max, _sparse_nan_min_max
 from ..utils.validation import _check_sample_weight
 from .sparsefuncs_fast import (
     csc_mean_variance_axis0 as _csc_mean_var_axis0,
@@ -410,72 +411,6 @@ def inplace_swap_column(X, m, n):
         inplace_swap_row_csc(X, m, n)
     else:
         _raise_typeerror(X)
-
-
-def _minor_reduce(X, ufunc):
-    major_index = np.flatnonzero(np.diff(X.indptr))
-
-    # reduceat tries casts X.indptr to intp, which errors
-    # if it is int64 on a 32 bit system.
-    # Reinitializing prevents this where possible, see #13737
-    X = type(X)((X.data, X.indices, X.indptr), shape=X.shape)
-    value = ufunc.reduceat(X.data, X.indptr[major_index])
-    return major_index, value
-
-
-def _min_or_max_axis(X, axis, min_or_max):
-    N = X.shape[axis]
-    if N == 0:
-        raise ValueError("zero-size array to reduction operation")
-    M = X.shape[1 - axis]
-    mat = X.tocsc() if axis == 0 else X.tocsr()
-    mat.sum_duplicates()
-    major_index, value = _minor_reduce(mat, min_or_max)
-    not_full = np.diff(mat.indptr)[major_index] < N
-    value[not_full] = min_or_max(value[not_full], 0)
-    mask = value != 0
-    major_index = np.compress(mask, major_index)
-    value = np.compress(mask, value)
-
-    if axis == 0:
-        res = sp.coo_matrix(
-            (value, (np.zeros(len(value)), major_index)), dtype=X.dtype, shape=(1, M)
-        )
-    else:
-        res = sp.coo_matrix(
-            (value, (major_index, np.zeros(len(value)))), dtype=X.dtype, shape=(M, 1)
-        )
-    return res.A.ravel()
-
-
-def _sparse_min_or_max(X, axis, min_or_max):
-    if axis is None:
-        if 0 in X.shape:
-            raise ValueError("zero-size array to reduction operation")
-        zero = X.dtype.type(0)
-        if X.nnz == 0:
-            return zero
-        m = min_or_max.reduce(X.data.ravel())
-        if X.nnz != np.prod(X.shape):
-            m = min_or_max(zero, m)
-        return m
-    if axis < 0:
-        axis += 2
-    if (axis == 0) or (axis == 1):
-        return _min_or_max_axis(X, axis, min_or_max)
-    else:
-        raise ValueError("invalid axis, use 0 for rows, or 1 for columns")
-
-
-def _sparse_min_max(X, axis):
-    return (
-        _sparse_min_or_max(X, axis, np.minimum),
-        _sparse_min_or_max(X, axis, np.maximum),
-    )
-
-
-def _sparse_nan_min_max(X, axis):
-    return (_sparse_min_or_max(X, axis, np.fmin), _sparse_min_or_max(X, axis, np.fmax))
 
 
 def min_max_axis(X, axis, ignore_nan=False):
